+++ whereguesthome/WebContent/WEB-INF/admin/product/list.jsp
@@ -193,7 +193,8 @@
 											</td>
 									
 											<td align="center" style="HEIGHT: 22px">
-												<a href="${g.gId}/delete">
+											<%-- 	<a href="${g.gId}/delete"> --%>
+											<a href="javascript: if(window.confirm('是否删除？')){window.location.href='${g.gId}/delete'}" >
 													<img src="${pageContext.request.contextPath}/images/i_del.gif" width="16" height="16" border="0" style="CURSOR: hand">
 												</a>
 											</td>
